requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
PyJWT==2.8.0
Pillow==10.1.0
cryptography==41.0.7
redis==5.0.1
rq==1.15.1
gunicorn==21.2.0
